chore: move components into playground
