+++ e2e/playground.spec.ts
@@ -1,22 +1,24 @@
 import { Page, expect, test } from '@playwright/test';
-import { PLAYGROUND_TAG } from '../src/client/components/playground/playgroundTags';
 import { isAgentResponse, isServerResponse } from './zodUtils';
 import { blockGoogleTagManager } from './e2eTestUtils';
+import { TEST_IDS } from '../src/client/testIDs';
+
+const TEST_ID = TEST_IDS.playground;
 
 const getAgentResponse = async (page: Page) => {
   const agentResponse =
-    (await (await page.getByTestId(PLAYGROUND_TAG.agentResponseJSON)).textContent()) ?? 'Agent response not found';
+    (await (await page.getByTestId(TEST_ID.agentResponseJSON)).textContent()) ?? 'Agent response not found';
   return JSON.parse(agentResponse);
 };
 
 const getServerResponse = async (page: Page) => {
   const serverResponse =
-    (await (await page.getByTestId(PLAYGROUND_TAG.serverResponseJSON)).textContent()) ?? 'Server response not found';
+    (await (await page.getByTestId(TEST_ID.serverResponseJSON)).textContent()) ?? 'Server response not found';
   return JSON.parse(serverResponse);
 };
 
 const clickPlaygroundRefreshButton = async (page: Page) => {
-  await page.getByTestId(PLAYGROUND_TAG.refreshButton).first().click();
+  await page.getByTestId(TEST_ID.refreshButton).first().click();
   // Artificial wait necessary to make sure you get the updated response every time
   await page.waitForTimeout(3000);
 };
@@ -30,7 +32,7 @@ test.describe('Playground page', () => {
   test('Page renders basic skeleton elements', async ({ page }) => {
     await page.getByText('Fingerprint Pro Playground', { exact: true }).waitFor();
     await page.getByText('Welcome, your visitor ID is').waitFor();
-    await page.getByTestId(PLAYGROUND_TAG.refreshButton).first().waitFor();
+    await page.getByTestId(TEST_ID.refreshButton).first().waitFor();
 
     await page.getByText('Identification', { exact: true }).waitFor();
     await page.getByText('Smart signals', { exact: true }).waitFor();
