+++ src/app/playground/components/BotDetectionResult.tsx
