+++ src/app/playground/components/VpnDetectionResult.tsx
